style(Testcase): 클래스 이름 변경
클래스 이름 앞에 Command 추가
