+++ project-template/src/main/java/org/teamone/projecttemplate/command/vo/CommandTestcaseResponse.java
@@ -1,6 +1,9 @@
 package org.teamone.projecttemplate.command.vo;
 
-public class TestcaseRequest {
+import lombok.Setter;
+
+@Setter
+public class CommandTestcaseResponse {
     private int testNo;
     private String separate;
     private String content;
@@ -9,17 +12,7 @@ public class TestcaseRequest {
     private String note;
     private int projectId;
 
-    public TestcaseRequest() {
-    }
-
-    public TestcaseRequest(int testNo, String separate, String content, String expectedValue, String result, String note, int projectId) {
-        this.testNo = testNo;
-        this.separate = separate;
-        this.content = content;
-        this.expectedValue = expectedValue;
-        this.result = result;
-        this.note = note;
-        this.projectId = projectId;
+    public CommandTestcaseResponse() {
     }
 
     public int getTestNo() {
@@ -50,9 +43,37 @@ public class TestcaseRequest {
         return projectId;
     }
 
+    public void setTestNo(int testNo) {
+        this.testNo = testNo;
+    }
+
+    public void setSeparate(String separate) {
+        this.separate = separate;
+    }
+
+    public void setContent(String content) {
+        this.content = content;
+    }
+
+    public void setExpectedValue(String expectedValue) {
+        this.expectedValue = expectedValue;
+    }
+
+    public void setResult(String result) {
+        this.result = result;
+    }
+
+    public void setNote(String note) {
+        this.note = note;
+    }
+
+    public void setProjectId(int projectId) {
+        this.projectId = projectId;
+    }
+
     @Override
     public String toString() {
-        return "TestcaseRequest{" +
+        return "TestcaseResponse{" +
                 "testNo=" + testNo +
                 ", separate='" + separate + '\'' +
                 ", content='" + content + '\'' +
